Added idle, move, jump, and fall animation settings to GenControl. Added GenTimer class
Added the ability to set separate idle, move, jump, and fall animations
in a GenControl object for simplicity. Added the ability to start or
stop an animation using IsPlaying, as well as control whether the
animation will loop or not using IsLooped. Added a new GenTimer class to
invoke a method after a given amount of time has passed.