Updated path following. Added moving platforms. Added movement helper methods.
Added more functionality to path following such as moving by a constant
velocity or acceleration. Added the Yoyo path following type. Objects
can now be set as platforms, so colliding objects that sit on top of a
platform object will move horizontally along with it. Added various
movement helper methods such as moving or accelerating an object towards
a point or towards the direction of a degree angle.