Added separate render modes. Added global volume control/display. Added game pausing. Fixed fade frame hang
The pixel render mode draws textures with sharp edges, and the smooth
render mode draws textures with smooth edges. Added a global volume
control that affects all sound volumes, with a volume control display
that appears when the +/- keys are pressed. Added game pausing to pause
updates for state members. Fixed the hanging frame issue when using a
camera fade to switch states, and the previous state would pop in for
one frame before switching to the new state.